+++ main/library_tree/initialise.js
@@ -1,9 +1,10 @@
 ﻿'use strict';
-//06/02/26
+//19/04/26
 
 /* global UserInterface:readable, Panel:readable, Scrollbar:readable, Vkeys:readable, Library:readable, Populate:readable, Search:readable, Find:readable, Buttons:readable, PopUpBox:readable, MenuItems:readable, Timers:readable */
+/* global require:readable */
 
-/* exported ui, panel, sbar, vk, lib, pop, search, but, find, popUpBox, men, timer */
+/* exported ui, panel, sbar, vk, lib, pop, search, but, find, popUpBox, men, timer, Chroma */
 
 /* eslint-disable no-redeclare */
 
@@ -21,3 +22,4 @@ but = new Buttons; // Regorxxx <- Filter / View / Source button ->
 const popUpBox = new PopUpBox;
 const men = new MenuItems;
 const timer = new Timers;
+const Chroma = require('..\\helpers-external\\chroma.js\\chroma.min'); // Relative to helpers folder